meeting add implemented, need to finish action
